+++ CHANGELOG.md
@@ -5,8 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
-
 ## Unreleased
+
+### Changed
+
+
+## v0.4.0 - 27.11.2021
 
 ### Added